backend/src/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-jose[cryptography]
bcrypt
passlib[bcrypt]
pydantic-settings
pydantic[email]
python-multipart